extract avatar into avatar.rs
